refactor: share event filter logic
