+++ src/pages/publicCalendar.tsx
@@ -16,7 +16,7 @@ import style1 from "../styles/calendar.module.css";
 import { useClerk } from "@clerk/clerk-react";
 import { EventDocument } from "../database/eventSchema";
 import { useRouter } from "next/router";
-import { convertEventDatesToDates } from "../utils/events";
+import { convertEventDatesToDates, filterEvents } from "../utils/events";
 import { DateTime } from "luxon";
 import { useSession } from "@clerk/nextjs";
 import { useUser } from "@clerk/clerk-react";
@@ -69,32 +69,13 @@ export default function CalendarPage() {
   const [toolbarSearchStyle, setToolbarSearchStyle] =
     useState<React.CSSProperties>({ display: "none" });
   const filteredEvents = useMemo(() => {
-    const normalizedSearchTerm = searchTerm.trim().toLowerCase();
-
-    return events.filter((event) => {
-      const passesOrganization = !hiddenOrganizations.includes(
-        event.organization,
-      );
-      const passesLocation = event.isVirtual ? showVirtual : showInPerson;
-
-      if (!passesOrganization || !passesLocation) {
-        return false;
-      }
-
-      if (!normalizedSearchTerm) {
-        return true;
-      }
-
-      const title = event.title?.toLowerCase() ?? "";
-      const description = event.description?.toLowerCase() ?? "";
-      const organization = event.organization?.toLowerCase() ?? "";
-
-      return (
-        title.includes(normalizedSearchTerm) ||
-        description.includes(normalizedSearchTerm) ||
-        organization.includes(normalizedSearchTerm)
-      );
-    });
+    return filterEvents(
+      events,
+      searchTerm,
+      hiddenOrganizations,
+      showVirtual,
+      showInPerson,
+    );
   }, [events, hiddenOrganizations, searchTerm, showInPerson, showVirtual]);
   const visibleMonthEvents = useMemo(() => {
     if (!visibleDateRange) {
